Move MarkdownTraverser to service (#12282)
